appendix_IV edits
I really like the idea of this appendix, but I question how well it's going to end up working out in print.  I wonder if it'd have a better life as only being part of the Chimera-version of the book.  Though as long as you keep it updated, POD will ensure future print versions get the updates anyway.