Tell the reader which html file this is
It took me a few moments of grepping back through
the book to figure out which html file this was
that we were editing.

This isn't needed if the filename in the markup surrounding
the source code somehow gets displayed as a caption for the
source code in the rendered book - but I don't see it in the
PDF version I'm looking at.