change order of last two points in Overview of Deployment
The last two bullet points appeared to be out of chronological order:
the one about deploying to production mentions the deployment script
for staging. So, the other list item (the one that mentions Ansible being
used for an automated script for deploying to staging) should come first.

I had to commit this change once again, because git ended up confused
after merging the `main` where both lines affected in this swap, were changed.