chapter9:optional: add --rm to `docker run busybox echo hello world`
This commit proposes to replace this command

  $ docker run busybox echo hello world

with

  $ docker run --rm busybox echo hello world

so that docker deletes the container when the container stops.

In this chapter, the docker commands never restart a stopped
container. Because of that, each `docker run` command could
use `--rm` to remove the container when the container stops.

Without `--rm`, the stopped containers accumulate.
`docker ps -a` shows the stopped containers not yet deleted.