chapter 9: format
replace

 $ *docker ps*
->
 $ docker ps

Before this commit, the rendered output is

  $ *docker ps

Notice the * in front of the command `docker`.

The edit in this commit is an attempt to remove that * from
before `docker`.

I see elsewhere in the source that uses * to bold text.
I have not investigated why the * here does not bold the text
and instead leaves the * character in the text, e.g. I have not
tried to use pass:quotes[] or [subs="specialcharacters,macros"].
Using one of those may be better than what I do in this commit.
I confess I have not investigated how to render the .asciidoc files
so that I can check the formatting.