chapter9: format *docker ps*
This commit adds '*'s and [subs="quotes"] to make bold shell commands
- docker ps
- docker kill
in listing blocks. This makes these shell commands appear
as other shell commands on the page do -- in bold in
listing blocks.

= Background

== I previously removed the '*'s and now I am adding them back. Why?
In an earlier commit (b65eb467d9fb1f22db98675549217dcb0c78889a) I removed '*'s
from *docker ps*.

I removed '*'s because the text I saw on
https://www.obeythetestinggoat.com/book/chapter_09_docker.html
included

  $ *docker ps

That is, '*docker' appeared on the rendered page.

I took it as given that the current version of the source code is
deployed on
https://www.obeythetestinggoat.com/book/chapter_09_docker.html.
[That conclusion was incorrect.]

Because I took it as given that the current version of the source is
deployed on the site, I thought there was a problem with

$ *docker ps*

in the current version of the source code. So, I removed the *'s.

== The reason for *docker on the site

Commit (c2b60ee10c2c19f08bf6e927c806fed1193dae6e) contains
```
$ *docker ps
CONTAINER ID
```
on lines 854-855. I hypothesize a version of the file with that in place
is what is deployed on the site and is the cause of what I saw on the
page.

Commit (97e9f6014587a353e2720709fb991b4e1fb4b38a) added a second star
```
$ *docker ps*
```
but this is not yet deployed (I believe).

== This commit adds [subs="quotes"] before the listing blocks for
`docker ps` and `docker kill`

The file contains format directives elsewhere that use
[subs="quotes"] before a block listing. Further, this seems
correct given what I read on these pages:
- https://docs.asciidoctor.org/asciidoc/latest/subs/apply-subs-to-blocks/
- https://docs.asciidoctor.org/asciidoc/latest/subs/quotes/

Alternatively,
[subs="macros"] with
$ pass:quotes[*docker ps*]
seems to work.

I read
https://github.com/hjwp/Book-TDD-Web-Dev-Python?tab=readme-ov-file#building-the-book-as-html
and built chapter_09_docker.html locally and it looks how I expect.
I will attach an image.