chapter 11: Getting our image onto the server: remove comment
This commit removes a comment.

26ea97a32b96791859e429407486c690506aa6b2 added a comment about an
error I encountered when I executed the `ansible-playbook ...` command
in section 'Getting our image onto the server'.

Xronophobe was able to reproduce the error only when the user
 - `pip install ansible` and
 - does not `pip install docker`

That is what I did. In the current version of the book at
[online](https://www.obeythetestinggoat.com/book/chapter_11_ansible.html#_installing_ansible)
I saw only the requirement to install ansible.
I failed to look at the current version of the code in the
[repository](https://github.com/hjwp/Book-TDD-Web-Dev-Python/blob/main/chapter_11_ansible.asciidoc?plain=1#L320-L322)
and see the additional requirement to install the docker python
package.

So, this commit removes that comment. Other users will not encounter
the error I did because they will both:
- pip install ansible, and
- pip install docker